requirements.txt
alembic==0.9.9
astroid==2.4.2
beautifulsoup4==4.9.3
blinker==1.4
blis==0.7.4
boilerpipe3==1.3
boilerpy3==1.0.3
breadability==0.1.20
bs4==0.0.1
catalogue==1.0.0
certifi==2020.12.5
chardet==3.0.4
click==6.7
colorama==0.4.4
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
flake8==3.8.4
Flask==1.0.2
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Material==0.1.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.2
gensim==3.8.3
gensim-sum-ext==0.1.2
gunicorn==20.0.4
httplib2==0.11.3
idna==2.6
isort==5.7.0
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
joblib==1.0.0
JPype1==1.2.1
kiwisolver==1.3.1
lazy==1.3
lazy-object-proxy==1.4.3
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
murmurhash==1.0.5
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.5
oauth2==1.9.0.post1
oauthlib==2.0.7
pandas==1.2.1
Pillow==8.1.0
plac==1.1.3
preshed==3.0.5
pycodestyle==2.6.0
pycorenlp==0.3.0
pycountry==20.7.3
pyflakes==2.2.0
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.3
pytz==2020.5
PyYAML==5.4.1
regex==2020.11.13
requests==2.18.4
requests-file==1.5.1
requests-oauthlib==0.8.0
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
smart-open==4.1.2
soupsieve==2.1
spacy==2.3.5
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.33.2
srsly==1.0.5
sumy==0.8.1
thinc==7.4.5
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tqdm==4.56.0
urllib3==1.22
URLObject==2.4.3
wasabi==0.8.0
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.1
